kyb_project/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-language==2.17.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
joblib==1.4.2
nltk==3.9.1
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.3
playwright==1.51.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyee==12.1.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.39
sqlparse==0.5.3
tenacity==9.0.0
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
